Move existing json converter into subdir
We're going to have a couple of these and they aren't really utility classes
anyway, so better off in their own folder.
